New verison in the works
This version is still in the works, it eliminates the need to manually create the folders before running the script.
I also started to build an other script that filters all of the IP's that send these requests, I can already extract them, next step is to transform these IPs I'm extracting now into a dictionary so that I can count how many requests each IP does.

This is still in very early development try it at your own risk, by the way the ip_filter script is not finalized yet so running it will most likely fail